+++ terraform/environments/production/locals.tf
@@ -4,4 +4,10 @@ locals {
     Environment = var.environment
     ManagedBy   = "Terraform"
   }
+
+  availability_zones = slice(
+    data.aws_availability_zones.available.names,
+    0,
+    2
+  )
 }