+++ terraform/modules/security-group/main.tf
@@ -1,0 +1,92 @@
+###########################################
+# Application Load Balancer Security Group
+###########################################
+
+resource "aws_security_group" "alb" {
+  name        = "${var.project_name}-alb-sg"
+  description = "Security group for the Application Load Balancer."
+  vpc_id      = var.vpc_id
+
+  tags = merge(
+    var.tags,
+    {
+      Name = "${var.project_name}-alb-sg"
+    }
+  )
+}
+
+resource "aws_vpc_security_group_ingress_rule" "alb_http" {
+  security_group_id = aws_security_group.alb.id
+
+  description = "Allow HTTP from the Internet"
+
+  ip_protocol = "tcp"
+
+  from_port = 80
+  to_port   = 80
+
+  cidr_ipv4 = "0.0.0.0/0"
+}
+
+resource "aws_vpc_security_group_ingress_rule" "alb_https" {
+  security_group_id = aws_security_group.alb.id
+
+  description = "Allow HTTPS from the Internet"
+
+  ip_protocol = "tcp"
+
+  from_port = 443
+  to_port   = 443
+
+  cidr_ipv4 = "0.0.0.0/0"
+}
+
+resource "aws_vpc_security_group_egress_rule" "alb_outbound" {
+  security_group_id = aws_security_group.alb.id
+
+  description = "Allow outbound traffic"
+
+  ip_protocol = "-1"
+
+  cidr_ipv4 = "0.0.0.0/0"
+}
+
+#########################################
+# EC2 Security Group
+#########################################
+
+resource "aws_security_group" "ec2" {
+  name        = "${var.project_name}-ec2-sg"
+  description = "Security group for EC2 instances."
+  vpc_id      = var.vpc_id
+
+  tags = merge(
+    var.tags,
+    {
+      Name = "${var.project_name}-ec2-sg"
+    }
+  )
+}
+
+resource "aws_vpc_security_group_ingress_rule" "ec2_http" {
+  security_group_id = aws_security_group.ec2.id
+
+  description = "Allow HTTP from the ALB"
+
+  ip_protocol = "tcp"
+
+  from_port = 80
+  to_port   = 80
+
+  referenced_security_group_id = aws_security_group.alb.id
+}
+
+resource "aws_vpc_security_group_egress_rule" "ec2_outbound" {
+  security_group_id = aws_security_group.alb.ec2.id
+
+  description = "Allow outbound traffic"
+
+  ip_protocol = "-1"
+
+  cidr_ipv4 = "0.0.0.0/0"
+}