+++ terraform/modules/security-group/outputs.tf
@@ -1,0 +1,11 @@
+output "alb_security_group_id" {
+  description = "ID of the Application Load Balancer security group."
+
+  value = aws_security_group.alb.id
+}
+
+output "ec2_security_group_id" {
+  description = "ID of the EC2 security group."
+
+  value = aws_security_group.ec2.id
+}